requirements.txt
pytest==8.1.1
selenium==4.19.0
allure-pytest==2.13.5